Collision Detection System Rework
Beginning to rework the collision detection system into something more
workable/modular. The current iteration is nice for a prototype, but I
need to give some serious TLC if I want to apply the system to anything
other than a prototype. This work will likely be applied in
mew2057/PenultEngine.